+++ libkineto/src/CuptiActivityApi.cpp
@@ -118,7 +118,7 @@ static bool nextActivityRecord(
   return record != nullptr;
 }
 
-void CuptiActivityApi::setMaxBufferSize(int size) {
+void CuptiActivityApi::setMaxBufferSize(int64_t size) {
   maxGpuBufferCount_ = 1 + size / kBufSize;
 }
